+++ plugins/flatlaf/src/main/java/is/codion/plugin/flatlaf/FlatLookAndFeelProvider.java
@@ -18,8 +18,10 @@
  */
 package is.codion.plugin.flatlaf;
 
+import is.codion.swing.common.ui.Utilities;
 import is.codion.swing.common.ui.laf.LookAndFeelEnabler;
 import is.codion.swing.common.ui.laf.LookAndFeelProvider;
+import is.codion.swing.common.ui.scaler.Scaler;
 
 import com.formdev.flatlaf.FlatDarculaLaf;
 import com.formdev.flatlaf.FlatDarkLaf;
@@ -28,29 +30,52 @@ import com.formdev.flatlaf.FlatLightLaf;
 import com.formdev.flatlaf.themes.FlatMacDarkLaf;
 import com.formdev.flatlaf.themes.FlatMacLightLaf;
 
+import javax.swing.UIManager;
 import javax.swing.UIManager.LookAndFeelInfo;
 import java.util.Collection;
+import java.util.function.Consumer;
 
 import static is.codion.swing.common.ui.laf.LookAndFeelEnabler.lookAndFeelEnabler;
 import static java.util.Arrays.asList;
 import static java.util.Collections.unmodifiableList;
+import static java.util.Objects.requireNonNull;
 
 /**
  * Provides all available Flat Look and Feels
  */
 public final class FlatLookAndFeelProvider implements LookAndFeelProvider {
 
+	private static final Consumer<LookAndFeelInfo> ENABLER = new DefaultEnabler();
+
 	public FlatLookAndFeelProvider() {}
 
 	@Override
 	public Collection<LookAndFeelEnabler> get() {
 		return unmodifiableList(asList(
-						lookAndFeelEnabler(new LookAndFeelInfo("Flat Darcula ", FlatDarculaLaf.class.getName())),
-						lookAndFeelEnabler(new LookAndFeelInfo("Flat Dark", FlatDarkLaf.class.getName())),
-						lookAndFeelEnabler(new LookAndFeelInfo("Flat IntelliJ", FlatIntelliJLaf.class.getName())),
-						lookAndFeelEnabler(new LookAndFeelInfo("Flat Light", FlatLightLaf.class.getName())),
-						lookAndFeelEnabler(new LookAndFeelInfo("Flat Mac Dark", FlatMacDarkLaf.class.getName())),
-						lookAndFeelEnabler(new LookAndFeelInfo("Flat Mac Light", FlatMacLightLaf.class.getName()))
+						lookAndFeelEnabler(new LookAndFeelInfo("Flat Darcula ", FlatDarculaLaf.class.getName()), ENABLER),
+						lookAndFeelEnabler(new LookAndFeelInfo("Flat Dark", FlatDarkLaf.class.getName()), ENABLER),
+						lookAndFeelEnabler(new LookAndFeelInfo("Flat IntelliJ", FlatIntelliJLaf.class.getName()), ENABLER),
+						lookAndFeelEnabler(new LookAndFeelInfo("Flat Light", FlatLightLaf.class.getName()), ENABLER),
+						lookAndFeelEnabler(new LookAndFeelInfo("Flat Mac Dark", FlatMacDarkLaf.class.getName()), ENABLER),
+						lookAndFeelEnabler(new LookAndFeelInfo("Flat Mac Light", FlatMacLightLaf.class.getName()), ENABLER)
 		));
+	}
+
+	/**
+	 * Requrired since scaling must happen before the look and feel is applied.
+	 */
+	private static final class DefaultEnabler implements Consumer<LookAndFeelInfo> {
+
+		@Override
+		public void accept(LookAndFeelInfo lookAndFeelInfo) {
+			try {
+				Scaler.instance(lookAndFeelInfo.getClassName()).ifPresent(Scaler::apply);
+				UIManager.setLookAndFeel(requireNonNull(lookAndFeelInfo).getClassName());
+				Utilities.updateComponentTreeForAllWindows();
+			}
+			catch (Exception e) {
+				throw new RuntimeException(e);
+			}
+		}
 	}
 }