+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/scaler/FontSizeScaler.java
@@ -16,51 +16,49 @@
  *
  * Copyright (c) 2025, Björn Darri Sigurðsson.
  */
-package is.codion.swing.common.ui.font;
-
-import is.codion.common.Configuration;
-import is.codion.common.property.PropertyValue;
-import is.codion.common.value.Value;
+package is.codion.swing.common.ui.scaler;
 
 import javax.swing.UIDefaults;
 import javax.swing.plaf.FontUIResource;
 import java.awt.Font;
 import java.util.Enumeration;
+import java.util.HashSet;
+import java.util.Set;
 
+import static java.util.Arrays.asList;
+import static java.util.Collections.unmodifiableSet;
+import static java.util.Objects.requireNonNull;
 import static javax.swing.UIManager.getLookAndFeelDefaults;
 
-/**
- * Provies a central location for font size configuration.
- */
-public final class FontSize {
+public final class FontSizeScaler implements Scaler {
 
-	private static final int DEFAULT_FONT_SIZE_RATIO = 100;
+	private static final Set<String> SUPPORTED_LOOK_AND_FEELS = unmodifiableSet(new HashSet<>(asList(
+					"javax.swing.plaf.metal.MetalLookAndFeel",
+					"javax.swing.plaf.nimbus.NimbusLookAndFeel",
+					"com.sun.java.swing.plaf.windows.WindowsLookAndFeel",
+					"com.sun.java.swing.plaf.gtk.GTKLookAndFeel",
+					"com.apple.laf.AquaLookAndFeel"
+	)));
 
-	/**
-	 * Specifies the global font size ratio.<br>
-	 * 85 = decrease the default font size by 15%<br>
-	 * 100 = use the default font size<br>
-	 * 125 = increase the default font size by 25%<br>
-	 * <p>Note that this does not support dynamic updates, application must be restarted for changes to take effect.
-	 * <ul>
-	 * <li>Value type: Integer
-	 * <li>Default value: 100
-	 * <li>Valid range: 50 - 200</li>
-	 * </ul>
-	 */
-	public static final PropertyValue<Integer> RATIO = Configuration.integerValue(FontSize.class.getName() + ".ratio", DEFAULT_FONT_SIZE_RATIO);
+	public FontSizeScaler() {}
 
-	static {
-		RATIO.addValidator(new FontSizeRatioValidator());
+	@Override
+	public void apply() {
+		if (RATIO.isNotEqualTo(DEFAULT_RATIO)) {
+			updateFontSize();
+		}
 	}
 
-	private FontSize() {}
+	@Override
+	public boolean supports(String lookAndFeelClassName) {
+		return SUPPORTED_LOOK_AND_FEELS.contains(requireNonNull(lookAndFeelClassName));
+	}
 
 	/**
 	 * Updates the font sizes for the current {@link UIDefaults} using the ratio from {@link #RATIO}.
-	 * Note that this must be done before the UI is initialized, dynamic font size changes are not supported.
+	 * Note that this must be done before the UI is initialized, dynamic scaling is not supported.
 	 */
-	public static void updateFontSize() {
+	private static void updateFontSize() {
 		UIDefaults defaults = getLookAndFeelDefaults();
 		float multiplier = RATIO.getOrThrow() / 100f;
 		Enumeration<Object> enumeration = defaults.keys();
@@ -76,16 +74,6 @@ public final class FontSize {
 				else {
 					defaults.put(key, derived);
 				}
-			}
-		}
-	}
-
-	private static final class FontSizeRatioValidator implements Value.Validator<Integer> {
-
-		@Override
-		public void validate(Integer value) {
-			if (value < 50 || value > 200) {
-				throw new IllegalStateException("Font size ratio must be between 50 and 200");
 			}
 		}
 	}